Add user and message queue interfaces
The user methods will handle current logged in users.  This will later
on be expanded to handle saving users to some permenant storage.  Since
logged in users will be synonymous with threads, there is a thread
handle method where the thread will spend most of its time.

The message queue interface is how the threads talk to each other.
There will be plenty of times that threads need to talk to each other,
for example if someone sends a 'tell' to another user or a simple
broadcast message.  Threads should exclusivly handle input and output to
their network socket, but anyone can write to their queue.